Paginacao e campos de busca de produto

+++ frontend/src/app/components/bandafrequencia/bandafrequencia-form/bandafrequencia-form.component.ts
@@ -10,6 +10,7 @@ import { MatToolbarModule } from '@angular/material/toolbar';
 import { ActivatedRoute, Router, RouterLink } from '@angular/router';
 import { BandaFrequencia } from '../../../models/banda-frequencia.model';
 import { BandaFrequenciaService } from '../../../services/banda-frequencia.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-bandafrequencia-form',
@@ -91,13 +92,35 @@ export class BandafrequenciaFormComponent {
 
   excluir() {
     const bandaFrequencia = this.formGroup.value;
-    this.bandaFrequenciaService.delete(bandaFrequencia).subscribe({
-      next: () => {
-        this.router.navigateByUrl('/admin/bandafrequencias');
-      },
-      error: (e) => {
-        console.log('Erro ao excluir', JSON.stringify(e));
-      },
+    Swal.fire({
+      title: 'Você tem certeza?',
+      text: 'Vou não vai poder reverter isso!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sim, deletar!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: 'Deletado!',
+          text: 'Banda de Frequência deletada com sucesso!',
+          icon: 'success',
+        });
+        
+        this.bandaFrequenciaService.delete(bandaFrequencia).subscribe({
+          next: () => {
+            this.router
+              .navigateByUrl('/', { skipLocationChange: true })
+              .then(() => {
+                this.router.navigate(['/admin/bandafrequencias']);
+              });
+          },
+          error: (e) => {
+            console.log('Erro ao excluir', JSON.stringify(e));
+          },
+        });
+      }
     });
   }
